fix: add check if room is free in dates before reserve confirm

+++ src/main/java/com/example/hostel1/servicies/logic/ReserveServiceImpl.java
@@ -11,6 +11,8 @@ import com.example.hostel1.repositories.spec.ReserveSpec;
 import com.example.hostel1.repositories.RoomRepository;
 import com.example.hostel1.servicies.ReserveService;
 import com.example.hostel1.servicies.exceptions.ResourceNotFoundException;
+import com.example.hostel1.servicies.exceptions.RoomIsExistException;
+import com.example.hostel1.servicies.exceptions.RoomIsNotFreeException;
 import com.example.hostel1.servicies.exceptions.ValidationException;
 import org.springframework.stereotype.Service;
 
@@ -42,19 +44,24 @@ public class ReserveServiceImpl extends GenericServiceImpl<Reserve> implements R
 
     @Transactional(rollbackOn = Exception.class)
     @Override
-    public Reserve createReserve(Order order) throws ResourceNotFoundException, ValidationException {
+    public Reserve createReserve(Order order) throws ResourceNotFoundException, ValidationException,
+            RoomIsNotFreeException {
         if (order != null) {
             Optional<Order> optional = orderRepository.findById(order.getId());
             if (optional.isPresent()) {
                 order = optional.get();
-                Reserve reserve = reserveFromOrder.reserveFromOrder(order);
-                reserve = reserveRepository.save(reserve);
-                orderRepository.delete(order);
-                return reserve;
+                if (isFreeInDates(order.getDateIn(), order.getDateOut(), order.getRoom().getId())) {
+                    Reserve reserve = reserveFromOrder.reserveFromOrder(order);
+                    reserve = reserveRepository.save(reserve);
+                    orderRepository.delete(order);
+                    return reserve;
+                }else throw new RoomIsNotFreeException("room is not free");
             } else throw new ResourceNotFoundException("order in not exist");
         } else throw new ValidationException("order must not be null");
     }
 
+
+    @Transactional(rollbackOn = Exception.class)
     @Override
     public List<Reserve> findByDatesIntervalAndRoom(
             Timestamp dateIn, Timestamp dateOut, Long roomId) throws ResourceNotFoundException, ValidationException {
